integrated generation of different formats in one goal
